Add task to run docker service before docker pull
Docker pull has been failed in circle ci due to Docker in Docker
problem. It needs `--privileged` option in additional, which will be
added in givery-technology/whale-env#34, but will work any environment.

I abandon to use service module in this task.
I'm not sure of the reason, but some how service module in ansible
doesn't work as I expected; it does not run docker service.